Consumer thread time statistics for signal processing
The goal is to have a way to quantify improvements in the signal
processing functions.

This change is not intended to be put to a production code, but
only be used during development.

The code waits for 1024 runs of the signal processing to happen
and after that prints timing information after every processing
run. In practice this means that one needs to wait a bit before
the timing information is printed.